+++ src/Subscriptions/SubscriptionBroadcaster.php
@@ -19,7 +19,7 @@ class SubscriptionBroadcaster implements BroadcastsSubscriptions
         protected AuthorizesSubscriptions $subscriptionAuthorizer,
         protected StoresSubscriptions $subscriptionStorage,
         protected SubscriptionIterator $subscriptionIterator,
-        protected BroadcastManager $broadcastManager,
+        protected BroadcastDriverManager $broadcastDriverManager,
         protected BusDispatcher $busDispatcher,
     ) {}
 
@@ -50,7 +50,7 @@ class SubscriptionBroadcaster implements BroadcastsSubscriptions
                     $subscriber->variables,
                     $subscriber,
                 );
-                $this->broadcastManager->broadcast($subscriber, $result);
+                $this->broadcastDriverManager->broadcast($subscriber, $result);
             },
         );
     }
@@ -58,7 +58,7 @@ class SubscriptionBroadcaster implements BroadcastsSubscriptions
     public function authorize(Request $request): Response
     {
         return $this->subscriptionAuthorizer->authorize($request)
-            ? $this->broadcastManager->authorized($request)
-            : $this->broadcastManager->unauthorized($request);
+            ? $this->broadcastDriverManager->authorized($request)
+            : $this->broadcastDriverManager->unauthorized($request);
     }
 }